move datalog headers to wpi/datalog
