+++ datalog/src/main/native/cpp/DataLogBackgroundWriter.cpp
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-#include <wpi/DataLogBackgroundWriter.h>
+#include <wpi/datalog/DataLogBackgroundWriter.h>
 
 #ifndef _WIN32
 #include <unistd.h>
