+++ datalog/src/main/native/include/wpi/datalog/DataLogBackgroundWriter.h
@@ -12,8 +12,8 @@
 #include <string_view>
 #include <thread>
 
-#include <wpi/DataLog.h>
 #include <wpi/condition_variable.h>
+#include <wpi/datalog/DataLog.h>
 #include <wpi/mutex.h>
 
 namespace wpi {
