+++ datalog/src/main/native/include/wpi/datalog/DataLogReaderThread.h
@@ -13,9 +13,9 @@
 #include <utility>
 #include <vector>
 
-#include <wpi/DataLogReader.h>
 #include <wpi/DenseMap.h>
 #include <wpi/Signal.h>
+#include <wpi/datalog/DataLogReader.h>
 #include <wpi/mutex.h>
 #include <wpi/struct/DynamicStruct.h>
