+++ datalog/src/main/native/include/wpi/datalog/FileLogger.h
@@ -8,7 +8,7 @@
 #include <string_view>
 #include <thread>
 
-#include <wpi/DataLog.h>
+#include <wpi/datalog/DataLog.h>
 
 namespace wpi {
 /**
