+++ datalog/src/test/native/cpp/DataLogTest.cpp
@@ -8,8 +8,8 @@
 #include <vector>
 
 #include <gtest/gtest.h>
-#include <wpi/DataLogWriter.h>
 #include <wpi/Logger.h>
+#include <wpi/datalog/DataLogWriter.h>
 #include <wpi/raw_ostream.h>
 
 namespace {
