+++ ntcore/src/main/native/cpp/local/LocalStorageImpl.cpp
@@ -11,7 +11,7 @@
 
 #include <fmt/format.h>
 #include <fmt/ranges.h>
-#include <wpi/DataLog.h>
+#include <wpi/datalog/DataLog.h>
 #include <wpi/SmallString.h>
 #include <wpi/StringExtras.h>
