+++ sysid/src/main/native/cpp/view/LogLoader.cpp
@@ -15,7 +15,7 @@
 #include <imgui.h>
 #include <imgui_stdlib.h>
 #include <portable-file-dialogs.h>
-#include <wpi/DataLogReaderThread.h>
+#include <wpi/datalog/DataLogReaderThread.h>
 #include <wpi/SmallVector.h>
 #include <wpi/SpanExtras.h>
 #include <wpi/StringExtras.h>
