+++ sysid/src/main/native/include/sysid/view/DataSelector.h
@@ -12,7 +12,7 @@
 #include <vector>
 
 #include <glass/View.h>
-#include <wpi/DataLogReaderThread.h>
+#include <wpi/datalog/DataLogReaderThread.h>
 #include <wpi/StringMap.h>
 
 #include "sysid/analysis/Storage.h"
