+++ wpilibc/src/main/native/cpp/DriverStation.cpp
@@ -23,7 +23,7 @@
 #include <networktables/NetworkTable.h>
 #include <networktables/NetworkTableInstance.h>
 #include <networktables/StringTopic.h>
-#include <wpi/DataLog.h>
+#include <wpi/datalog/DataLog.h>
 #include <wpi/EventVector.h>
 #include <wpi/condition_variable.h>
 #include <wpi/json.h>
